LevelScript::m_bStartSequence is uint8 not uint16 — puts m_bStartSequenceOver at 0x21 — +1/+1
ASM (resetStartSequenceOver): original writes m_bStartSequenceOver at [r0,#0x21]; ours had it at 0x22 because
m_bStartSequence was uint16. It's a "b" boolean flag — changed to uint8. Fixes offset + store order.
linked_exact 1957->1958, byte_exact 1094->1095.

+++ src/game/world/LevelScript.h
@@ -25,7 +25,7 @@ public:
     TargetFollowCamera *m_pCamera;
     Level *m_pLevel;
     int m_nState;
-    uint16_t m_bStartSequence;
+    uint8_t m_bStartSequence; // ASM (resetStartSequenceOver): m_bStartSequenceOver is at 0x21, so this is 1 byte
     uint8_t m_bStartSequenceOver;
     int field_0x24;
     int field_0x28;
